refactor: copy-monthを選択フックと項目グループに分割
