+++ src/features/copy-month/types.ts
@@ -1,0 +1,2 @@
+// 項目の選択状態（未選択 | 金額込み | 項目名のみ）
+export type ItemSelection = 'none' | 'withAmount' | 'labelOnly'
